backend/requirements.txt
fastapi
uvicorn
firecrawl-py
langchain-google-genai
pydantic
python-dotenv
langchain
